refactor: update import paths for timeline store files and introduce new base timeline store
