+++ apps/web/ce/store/root.store.ts
@@ -6,8 +6,8 @@
 
 // store
 import { CoreRootStore } from "@/store/root.store";
-import type { ITimelineStore } from "./timeline";
-import { TimeLineStore } from "./timeline";
+import type { ITimelineStore } from "@/store/timeline/timeline.store";
+import { TimeLineStore } from "@/store/timeline/timeline.store";
 
 export class RootStore extends CoreRootStore {
   timelineStore: ITimelineStore;
