+++ apps/web/core/store/timeline/issues-timeline.store.ts
@@ -5,10 +5,9 @@
  */
 
 import { autorun } from "mobx";
-// Plane-web
-import type { RootStore } from "@/plane-web/store/root.store";
-import type { IBaseTimelineStore } from "@/plane-web/store/timeline/base-timeline.store";
-import { BaseTimeLineStore } from "@/plane-web/store/timeline/base-timeline.store";
+import type { CoreRootStore as RootStore } from "@/store/root.store";
+import type { IBaseTimelineStore } from "@/store/timeline/base-timeline.store";
+import { BaseTimeLineStore } from "@/store/timeline/base-timeline.store";
 
 export interface IIssuesTimeLineStore extends IBaseTimelineStore {
   isDependencyEnabled: boolean;
