+++ apps/web/core/store/timeline/modules-timeline.store.ts
@@ -5,10 +5,9 @@
  */
 
 import { autorun } from "mobx";
-// Store
-import type { RootStore } from "@/plane-web/store/root.store";
-import { BaseTimeLineStore } from "@/plane-web/store/timeline/base-timeline.store";
-import type { IBaseTimelineStore } from "@/plane-web/store/timeline/base-timeline.store";
+import type { CoreRootStore as RootStore } from "@/store/root.store";
+import { BaseTimeLineStore } from "@/store/timeline/base-timeline.store";
+import type { IBaseTimelineStore } from "@/store/timeline/base-timeline.store";
 
 export interface IModulesTimeLineStore extends IBaseTimelineStore {
   isDependencyEnabled: boolean;
